Revert "tweak(build): less compiler warnings in gta-net"
Uncertain reports that might indicate canary being net-broken.

Likely one of the operator precedence things being incorrectly assumed, or some other signed/unsigned behavior being explicitly wrong now.

This reverts commit 538f201466bbb22a549a114d86647e3d9cff5f19.

+++ code/client/shared/atPool.h
@@ -22,7 +22,7 @@ private:
 
 public:
 	template<typename T>
-	T* GetAt(size_t index) const
+	T* GetAt(int index) const
 	{
 		if (m_flags[index] < 0)
 		{
@@ -74,7 +74,7 @@ public:
 
 	void Clear()
 	{
-		for (size_t i = 0; i < m_count; i++)
+		for (int i = 0; i < m_count; i++)
 		{
 			if (m_flags[i] >= 0)
 			{
